requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==3.0.2
Django==4.2.13
django-cors-headers==4.3.1
django-eventstream==4.5.1
django-grip==3.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gripcontrol==4.2.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
pubcontrol==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygtail==0.14.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
requests==2.31.0
service-identity==24.1.0
shortuuid==1.0.13
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wgconfig==1.0.1
whitenoise==6.6.0
zope.interface==6.4.post2
